requirements.txt
paramiko==2.2.1
psutil==5.2.2
pycrypto==2.6.1